requirements.txt
streamlit
numpy
pandas
scikit-learn
pdfminer.six
python-docx
nltk
sqlalchemy
pycryptodome
reportlab
PyPDF2
